Fix #50, Separate config files
