+++ config/default_hk_msgids.h
