+++ config/default_hk_tbldefs.h
@@ -1,0 +1,40 @@
+/************************************************************************
+ * NASA Docket No. GSC-18,919-1, and identified as “Core Flight
+ * System (cFS) Housekeeping (HK) Application version 2.5.1”
+ *
+ * Copyright (c) 2021 United States Government as represented by the
+ * Administrator of the National Aeronautics and Space Administration.
+ * All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License. You may obtain
+ * a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ ************************************************************************/
+
+/**
+ * @file
+ *   Specification for the CFS Housekeeping (HK) table related
+ *   constant definitions.
+ */
+
+#ifndef HK_TBLDEFS_H
+#define HK_TBLDEFS_H
+
+/**
+ *  \brief Maximum Number of HK Copy Table Entries
+ *
+ *  \par Description:
+ *       Dictates the number of elements in the hk copy table.
+ *
+ *  \par Limits
+ *       The maximum size of this parameter is 8192
+ */
+#define HK_COPY_TABLE_ENTRIES 128
+
+#endif
