+++ docs/dox_src/hk-common.doxyfile.in
@@ -5,3 +5,4 @@
 # Include front material followed by everything in fsw
 INPUT      += @hk_MISSION_DIR@/docs/dox_src/cfs_hk.dox
 INPUT      += @hk_MISSION_DIR@/fsw
+INPUT      += @hk_MISSION_DIR@/config
