+++ fsw/src/hk_utils.h
@@ -28,7 +28,7 @@
  * Includes
  ************************************************************************/
 #include "cfe.h"
-#include "hk_tbldefs.h"
+#include "hk_tbl.h"
 
 /*************************************************************************
  * Macro definitions
