+++ unit-test/utilities/hk_test_utils.c
@@ -25,7 +25,7 @@
 #include "cfe_msgids.h"
 #include "hk_app.h"
 #include "hk_msg.h"
-#include "hk_events.h"
+#include "hk_eventids.h"
 #include "hk_version.h"
 #include "hk_msgids.h"
 #include "hk_test_utils.h"
