+++ unit-test/utilities/hk_test_utils.h
@@ -22,10 +22,10 @@
 
 #include "hk_app.h"
 #include "hk_msg.h"
-#include "hk_events.h"
+#include "hk_eventids.h"
 #include "hk_version.h"
 #include "hk_msgids.h"
-#include "hk_tbldefs.h"
+#include "hk_tbl.h"
 
 #include "cfe_msgids.h"
